Add Architecture Cookie with Detail, Search and Shared

+++ src/app/modules/rutas/detalle-rutas/detalle-rutas.component.ts
@@ -1,11 +1,13 @@
 import { Component, OnDestroy,  } from '@angular/core';
-import { ActivatedRoute} from '@angular/router';
+import { ActivatedRoute, Router} from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
 import { Subscription } from 'rxjs';
-import { COOKIE_ROUTE, StartRouteCookie } from 'src/app/core/data/cookie/start-cookie.interface';
+import { COOKIE_ROUTE, RutaCookie, StartRouteCookie } from 'src/app/core/data/cookie/start-cookie.interface';
 import { Ruta } from 'src/app/core/data/ruta.interface';
+import { SwitchService } from 'src/app/core/services/modal/switch.service';
 import { RouteService } from 'src/app/core/services/route/route.services';
 import { NavbarCommunicationService } from 'src/app/shared/services/navbar.service';
+import { StickButtonCommunicationService } from 'src/app/shared/services/stick-button.service';
 
 @Component({
   selector: 'app-detalle-rutas',
@@ -17,11 +19,16 @@ export class DetalleRutasComponent implements OnDestroy {
   public rutaCompleta: string = '';
   public existRuta: boolean = true;
   public cookie: StartRouteCookie;
+  public showModal: boolean = false;
   private suscriptionRutaService: Subscription;
+  private subscriptionSwitch: Subscription;
   constructor(private readonly route: ActivatedRoute, 
               private readonly rutaService: RouteService, 
               private readonly navbarService: NavbarCommunicationService,
-              private readonly cookieService: CookieService) {
+              private readonly cookieService: CookieService,
+              private readonly modalService: SwitchService,
+              private readonly router: Router,
+              private readonly stickButtonCommunicationService: StickButtonCommunicationService) {
     
     const id: number = this.route.snapshot.params['id'];
     this.navbarService.setActiveSearch(false);
@@ -34,11 +41,12 @@ export class DetalleRutasComponent implements OnDestroy {
                                                   this.ruta = ruta;
                                                   this.existRuta = this.ruta != null;
                                               });
-
+    this.subscriptionSwitch = this.modalService.$modal.subscribe((showModal: boolean) => this.showModal = showModal);
   }
 
   ngOnDestroy(): void {
     this.suscriptionRutaService.unsubscribe();
+    this.subscriptionSwitch.unsubscribe();
   }
   
   calcularDuracion(duracionTotal: number): string {
@@ -50,16 +58,23 @@ export class DetalleRutasComponent implements OnDestroy {
   }
 
   startRoute(): void {
-    if(this.ruta != null){
-      let start: StartRouteCookie = { ruta: this.ruta, init: true };
-      if (start.ruta != null) {
-        start.ruta.image = "";
+    if(this.cookie.init) this.showModal = true;
+    else  {
+      if (this.ruta != null) {
+        const { id, name, ubicacion, origenLatitud, origenLongitud, destinoLatitud, destinoLongitud } = this.ruta;
+        const buildRutaCookie: RutaCookie = { id, name, ubicacion, origenLatitud, origenLongitud, destinoLatitud, destinoLongitud };
+        const start: StartRouteCookie = { ruta: buildRutaCookie, init: true };
+        this.cookieService.delete(COOKIE_ROUTE);
+        this.cookieService.set(COOKIE_ROUTE, JSON.stringify(start));
+        this.router.navigate(['/ruta/start']);
       }
-      
-      this.cookieService.delete(COOKIE_ROUTE);
-      this.cookieService.set(COOKIE_ROUTE, JSON.stringify(start));
-      console.log(this.cookieService.get(JSON.stringify(start)))
     }
+  }
+
+  endRoute(): void {
+    this.cookieService.delete(COOKIE_ROUTE);
+    this.cookie = { ruta : null, init: false };
+    this.stickButtonCommunicationService.setActiveSticky(false);
   }
 
 }
